refactor(notification_service): move mongo related codes to independent files
